incorporates additional review feedback
* Removes commented field
* Updates copyright notice years
* consolidates external-facing logic in pgadmin package
