+++ internal/pgadmin/server.go
@@ -1,7 +1,7 @@
 package pgadmin
 
 /*
- Copyright 2017 - 2020 Crunchy Data Solutions, Inc.
+ Copyright 2020 Crunchy Data Solutions, Inc.
  Licensed under the Apache License, Version 2.0 (the "License");
  you may not use this file except in compliance with the License.
  You may obtain a copy of the License at
